CAMBIOSANA201024ESTADOCIVILESTADORECETA

+++ src/main/java/com/sisu/sisu/Dao/ITiposEstadoCivilDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.TiposEstadoCivil;
 
-public interface ITiposEstadoCivilDao extends CrudRepository<TiposEstadoCivil, Integer> {
+public interface ITiposEstadoCivilDao extends CrudRepository<TiposEstadoCivil, Long> {
 
 }
